+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.0.8 (2022-07-29)
+
 ### Added
 
 -   `formSubmissionWorkflowEvents` and `formSubmissionSchedulingBooking` to `FormSubmissionMetadataResponse`